Phase3a実装: UI表示制御分離とdisplay.js作成
- script.js 107行削減（518行→411行）
- display.js新規作成（UI表示関数5個を純粋関数化）
- 168テスト100%通過維持
- MVC View層の確立
